_CSV table extended

+++ tasks/csv/src/controller.cpp
@@ -175,7 +175,7 @@ namespace node
 			boost::uuids::random_generator uidgen;
 
 			boost::random::mt19937 rng_;
-            rng_.seed(std::time(nullptr));
+            rng_.seed(static_cast<std::uint32_t>(std::time(nullptr)));
 			boost::random::uniform_int_distribution<int> monitor_dist(10, 120);
 
 			const auto conf = cyng::vector_factory({
